LayoutTests/fast/frames/seamless/seamless-designMode-expected.txt

Test that being a seamless iframe does not break designMode.
PASS window.iframe.contentDocument.body.textContent is "PASS"
